DEVX-1871 microsvcs: major refactors of local and docker demos
The general purpose of this change is to reduce duplication, standardize some usage
in docker-compose, and make some functions avaialable for use targeting other platforms
(ccloud).

Additionally, this change moves some sqlite code out of the shared utils folder and into
the microservices folder.  It changes the behavior of writing the database file to a
system level folder preferring modified files in local folders.

Connector configs now share a single template and are rendered to a local folder before
being posted to a connector cluster. This should reduce duplciation as well as allow for
posting from ccloud shortly.